refactor(application-admin-console): extract CreateIdpModal from IdentityProviders
Mirror of the admin-console split: IdentityProviders.tsx held the IdP list page
plus a self-contained CreateIdpModal, pulling 17 imports (over the high-coupling
threshold of 16). Extracting the modal to CreateIdpModal.tsx drops the page to
15 imports (off the high-coupling list).

With both apps' modals now isolated, the genuine cross-app duplication is
explicit and ready to dedup: this CreateIdpModal and the admin-console one are
~95% identical, differing only in the injected client type (TenantIdpClient vs
IdpClient) and error helper — a clean follow-up to extract into a shared,
client-parameterized component.

This completes the high-coupling cleanup for every cleanly-splittable page
(tabs / Home / ScoreEvents / AuditLog / IdentityProviders x2).

Relates #1418.

## Regression analysis
- Pure structural move: CreateIdpModal's body + props are byte-identical, only
  relocated and imported back. typecheck passes; all 407 application-admin-console
  tests pass unchanged. describeTenantIdpError stays used by the page.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
